refactor(app.js, auth.js): Extract all auth method to individual modules
